Extract connection + studio ops into host/{connection,studio}.ts
Move getVersion/exportState/importState/saveBranch/listBranches/switchBranch/
deleteBranch and the instance-id + branch-registry helpers into
host/connection.ts; move setPolicy/getPolicy/getSnapshot into host/studio.ts.
handleOp is now a pure orchestrator: the eager db/lens preamble followed by
family routing, then the unknown-method default.

host.ts re-exports INSTANCE_ID_KEY/randomUuid/getOrCreateInstanceId/
BRANCH_PREFIX/BRANCH_REGISTRY_KEY/listBranchNames from host/connection.js, so
the public export surface is unchanged (verified identical). Bodies verbatim;
orphaned pyric/sandbox bundle imports dropped from host.ts. No behavior change.
