+++ packages/pyric-tools/src/serve/worker/host.ts
@@ -54,8 +54,6 @@ import {
   type CollectionReference,
   type Query,
 } from 'pyric/firestore';
-import type { Sandbox, PersistenceBackend } from 'pyric/sandbox';
-import { serializeToBuckets, bundleRecords, parseBundle, deserializeFromBuckets } from 'pyric/sandbox';
 import {
   ref as rtdbRef,
   onValue as rtdbOnValue,
@@ -119,66 +117,22 @@ import { isRulesOp, handleRulesOp } from './host/rules.js';
 import { isAdminFirestoreOp, handleAdminFirestoreOp } from './host/admin-firestore.js';
 import { isRtdbOp, handleRtdbOp, rtdbSnapToWire } from './host/rtdb.js';
 import { isStorageOp, handleStorageOp } from './host/storage.js';
+import { isConnectionOp, handleConnectionOp } from './host/connection.js';
+import { isStudioOp, handleStudioOp } from './host/studio.js';
 
 // Re-export so host.ts's public surface is unchanged after the decomposition.
 export { ensureAuth, portSession } from './host-auth.js';
 export type { HostCtx, PortLike } from './host-context.js';
-
-/** Build hash injected by the bundler's esbuild `define`. Undefined when the
- *  compiled host is imported directly (tests) — guarded with `typeof`. */
-declare const __PYRIC_WORKER_VERSION__: string;
-
-/**
- * Per-SharedWorker instance id — generated once and persisted to the RAW idb
- * (local-only, like the session record above; it must NEVER reach the
- * committable server file). Because IndexedDB is per (origin + browser profile),
- * two profiles on the same `localhost:<port>` get two distinct ids — which is
- * exactly how the UI tells same-port-different-profile sandboxes apart.
- */
-export const INSTANCE_ID_KEY = 'pyric:worker:instance';
-
-/**
- * `crypto.randomUUID()` is secure-context-only (https or localhost), so it is
- * `undefined` over plain http on a non-localhost host (a Tailscale or LAN
- * hostname). `crypto.getRandomValues` is NOT gated, so build a v4 UUID from it as
- * the fallback. Without this the worker throws on init over Tailscale and the
- * whole sandbox (auth, firestore, bridge) silently fails to come up.
- */
-export function randomUuid(): string {
-  if (typeof crypto.randomUUID === 'function') return crypto.randomUUID();
-  const b = crypto.getRandomValues(new Uint8Array(16));
-  b[6] = (b[6] & 0x0f) | 0x40;
-  b[8] = (b[8] & 0x3f) | 0x80;
-  const h = Array.from(b, (x) => x.toString(16).padStart(2, '0'));
-  return `${h.slice(0, 4).join('')}-${h.slice(4, 6).join('')}-${h.slice(6, 8).join('')}-${h.slice(8, 10).join('')}-${h.slice(10, 16).join('')}`;
-}
-
-export async function getOrCreateInstanceId(idb: PersistenceBackend): Promise<string> {
-  const rec = (await idb.getRecord(INSTANCE_ID_KEY, 'id')) as { value?: string } | undefined;
-  if (rec && typeof rec.value === 'string') return rec.value;
-  const id = randomUuid();
-  await idb.putRecords(INSTANCE_ID_KEY, new Map([['id', { value: id }]]));
-  return id;
-}
-
-// ── Phase 3: named branches ─────────────────────────────────────────────────
-// A branch is a named saved state bundle in the RAW idb (local-only, like the
-// instance id + session; it must NEVER reach the committable server file). They
-// let one instance keep several named states it can switch between
-// (switchBranch = loadSnapshot the bundle, a clobber). A registry record holds
-// the ordered name list, since the backend lists records WITHIN a key, not keys.
-export const BRANCH_PREFIX = 'pyric:worker:branch:';
-export const BRANCH_REGISTRY_KEY = 'pyric:worker:branches';
-
-export async function listBranchNames(idb?: PersistenceBackend): Promise<string[]> {
-  if (!idb) return [];
-  const rec = (await idb.getRecord(BRANCH_REGISTRY_KEY, 'names')) as { value?: string[] } | undefined;
-  return Array.isArray(rec?.value) ? rec.value : [];
-}
-
-async function writeBranchRegistry(idb: PersistenceBackend, names: string[]): Promise<void> {
-  await idb.putRecords(BRANCH_REGISTRY_KEY, new Map([['names', { value: names }]]));
-}
+// Instance-id + named-branch surface (part of the host's public shape:
+// serve-init imports getOrCreateInstanceId; tests import the rest).
+export {
+  INSTANCE_ID_KEY,
+  randomUuid,
+  getOrCreateInstanceId,
+  BRANCH_PREFIX,
+  BRANCH_REGISTRY_KEY,
+  listBranchNames,
+} from './host/connection.js';
 
 // ─── Op handlers ──────────────────────────────────────────────────────────
 
@@ -202,118 +156,13 @@ async function handleOp(ctx: HostCtx, port: PortLike, msg: OpMessage): Promise<v
   if (isAdminFirestoreOp(msg.method)) return handleAdminFirestoreOp(ctx, port, msg);
   if (isRtdbOp(msg.method)) return handleRtdbOp(ctx, port, msg);
   if (isStorageOp(msg.method)) return handleStorageOp(ctx, port, msg);
+  if (isConnectionOp(msg.method)) return handleConnectionOp(ctx, port, msg);
+  if (isStudioOp(msg.method)) return handleStudioOp(ctx, port, msg);
 
-  switch (msg.method) {
-    case 'getVersion': {
-      // The build hash is injected by the bundler (esbuild `define`). `typeof`
-      // guards the non-bundled path (tests import the compiled host directly,
-      // where the global is undefined) → reports 'dev'.
-      ok(port, msg.id, {
-        version: typeof __PYRIC_WORKER_VERSION__ !== 'undefined' ? __PYRIC_WORKER_VERSION__ : 'dev',
-        instanceId: ctx.instanceId,
-      });
-      break;
-    }
-
-    case 'exportState': {
-      // Phase 2 (transfer): serialize the FULL sandbox state to a portable bundle
-      // string using the SAME chunk format the persist layer uses, so wrapper
-      // types (Timestamp / Bytes / GeoPoint / VectorValue) round-trip. The string
-      // crosses the MessagePort cleanly (unlike the raw snapshot object).
-      const snap = ctx.sandbox.snapshot();
-      ok(port, msg.id, { bundle: bundleRecords(serializeToBuckets(snap.firestore, snap.services, 0)) });
-      break;
-    }
-
-    case 'importState': {
-      // Phase 2 (clobber): replace this sandbox's ENTIRE state with the imported
-      // bundle via the public loadSnapshot() (reset + rebuild firestore + restore
-      // services; listeners re-evaluate, persist re-flushes).
-      ctx.sandbox.loadSnapshot(deserializeFromBuckets(parseBundle(msg.bundle)));
-      ok(port, msg.id, { ok: true });
-      break;
-    }
-
-    case 'saveBranch': {
-      // Phase 3: snapshot the live sandbox into a named branch bundle (raw idb).
-      if (!ctx.sessionBackend) {
-        ok(port, msg.id, { ok: false, error: 'no persistence backend' });
-        break;
-      }
-      const snap = ctx.sandbox.snapshot();
-      const bundle = bundleRecords(serializeToBuckets(snap.firestore, snap.services, 0));
-      await ctx.sessionBackend.putRecords(BRANCH_PREFIX + msg.name, new Map([['bundle', { value: bundle }]]));
-      const names = await listBranchNames(ctx.sessionBackend);
-      if (!names.includes(msg.name)) await writeBranchRegistry(ctx.sessionBackend, [...names, msg.name]);
-      ok(port, msg.id, { ok: true });
-      break;
-    }
-
-    case 'listBranches': {
-      ok(port, msg.id, { branches: await listBranchNames(ctx.sessionBackend) });
-      break;
-    }
-
-    case 'switchBranch': {
-      // Phase 3: loadSnapshot the named branch bundle (a clobber).
-      const rec = ctx.sessionBackend
-        ? ((await ctx.sessionBackend.getRecord(BRANCH_PREFIX + msg.name, 'bundle')) as { value?: string } | undefined)
-        : undefined;
-      if (!rec?.value) {
-        ok(port, msg.id, { ok: false, error: `no such branch: ${msg.name}` });
-        break;
-      }
-      ctx.sandbox.loadSnapshot(deserializeFromBuckets(parseBundle(rec.value)));
-      ok(port, msg.id, { ok: true });
-      break;
-    }
-
-    case 'deleteBranch': {
-      if (ctx.sessionBackend) {
-        await ctx.sessionBackend.clear(BRANCH_PREFIX + msg.name);
-        await writeBranchRegistry(
-          ctx.sessionBackend,
-          (await listBranchNames(ctx.sessionBackend)).filter((n) => n !== msg.name),
-        );
-      }
-      ok(port, msg.id, { ok: true });
-      break;
-    }
-
-    case 'setPolicy': {
-      // Store the dial's PolicyRequest as the worker-side runtime governance
-      // (Pyric Studio F3). This is the source of truth Studio reflects + a
-      // future in-worker agent runtime consults. It does NOT push into a
-      // running bridge process — see the limitation note on `ctx.policy` /
-      // `PolicyRequest`. Additive + idempotent: last write wins.
-      ctx.policy = msg.policy;
-      ok(port, msg.id, null);
-      break;
-    }
-
-    case 'getPolicy': {
-      // Read back the active runtime policy (null until the dial set one), so
-      // Studio can reflect persisted state across reconnects within a worker
-      // lifetime and a freshly-connecting port can hydrate the dial.
-      ok(port, msg.id, ctx.policy ?? null);
-      break;
-    }
-
-    case 'getSnapshot': {
-      // Export the current sandbox snapshot (Pyric Studio rules re-run): Studio
-      // forks it locally to test edited rules / re-issue as the user on a branch.
-      ok(port, msg.id, ctx.sandbox.snapshot());
-      break;
-    }
-
-    default: {
-      // Auth ops (`auth.*`) and AI ops (`ai.*`) are routed to handleAuthOp /
-      // handleAiOp by handleMessage before reaching here, so any method
-      // landing in this default is genuinely unknown. (We can't use a `never` exhaustiveness check anymore because
-      // OpMessage now includes the auth variants this switch deliberately skips.)
-      fail(port, msg.id, new Error(`Unknown method: ${String((msg as { method: unknown }).method)}`));
-    }
-  }
+  // Auth (`auth.*`), AI (`ai.*`), and messaging (`messaging.*`) ops are routed
+  // to their handlers by dispatchMessage BEFORE reaching handleOp, so any
+  // method landing here is genuinely unknown.
+  fail(port, msg.id, new Error(`Unknown method: ${String((msg as { method: unknown }).method)}`));
 }
 
 // ─── Subscription handler ─────────────────────────────────────────────────
